Ingestion_custom_py/requirements.txt
google-cloud-pubsub==2.20.0
google-cloud-bigquery==3.13.0